Started inventory merge
