+++ Assets/Prefabs/Environment.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 19ce7f3578779dc4ebec3f2d9f5d4ca7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
