+++ Assets/Prefabs/Environment/InteractableObject.prefab.meta
