+++ Assets/Prefabs/Managers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7ad724f6e5e1c4a438c4690f20c5adbc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
