+++ Assets/Scenes/DevScenes/roomStateSceneTest.unity
@@ -599,7 +599,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!114 &1684002571
 MonoBehaviour:
   m_ObjectHideFlags: 0
