--- Assets/Scripts/UI/Pause.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 73079be1433fc4845afdfc430468942d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
